Removed Newtonsoft.Json dependency in EventData.cs. Instead I changed it to use a RestSharp request and deserialize the objects directly to the classes defined in the new APIClasses.cs file. Added caching of event_names list so we don't repeatedly request that data.
This is just an example of another (and imo cleaner) way of doing this. If you don't like it, then just toss these changes in the trash. If you do like it, we can go through the other classes to make them work the same way.

Oh, I also added NUnit to the test project. Not sure why it wasn't there when I synced the solution down because it was clearly referencing it already.

git-tfs-id: [https://tfs.codeplex.com:443/tfs/TFS35]$/gw2dotnet/GW2.NET;C27949