Brought the code style in line with StyleCop code formatting rules.
Refctored the api call out of the EventData class into the ApiCall class (name is CallApi<T>). This made the code slightly longer (two lines) but reduced maintenance cost because the whole call plus deserialisation are in the API call class (we could also do the JSON.Net deserialisation there so we don't need the proxy classes and still don't have a reference in the other classes. Discussion about that go in the discussion topic on the codeplex site please).

git-tfs-id: [https://tfs.codeplex.com:443/tfs/TFS35]$/gw2dotnet/GW2.NET;C28004