Rewrote the object model. New version resides in the Gw2DotNew.V1 namespace. World Data 90% complete, only the GetEvent function is missing. Not totally happy with the names solution as it required a List<Dictionary<string,string>> to convert the data, will try to find a cleaner solution.
Unit test ran fine (even though it's just a barebones one) and matches the events propery. Now the user requests the Events property on the WorldManager and he get's the discovered events with name.

Next Step: Clean up code, but I'm too tired for that now.


Also it seems some events have the status invalid, which is akward since ANet never said anything about that. Included the status in the GwEventState class.

Please test the code and discuss it in the apropiate topic.

git-tfs-id: [https://tfs.codeplex.com:443/tfs/TFS35]$/gw2dotnet/GW2.NET;C28043