Moved all the implementation of the Data classes out of WorldManager and into those classes. Implemented IList<T> interface on those classes to achieve this.
Removed the Data class instantiation out of the WorldManager constructor and into properties so that we do not instantiate them until we actually need them. This also allows us to easily clear cached data by nulling out those members when the language is changed, and knowing that the next time they are accessed we'll automatically get new instances.

git-tfs-id: [https://tfs.codeplex.com:443/tfs/TFS35]$/gw2dotnet/GW2.NET;C28048